+++ CHANGELOG.md
@@ -6,9 +6,9 @@
 * Add `customer_device_id`, `customer_location_zip`, and `customer_tenure` to `RiskData`
 * Add `phone_number` and `shipping_method` to `Address`
 * Add validation error codes:
-  * `Address.ShippingMethodIsInvalid`
-  * `Address.ShippingPhoneNumberIsInvalid`
-  * `Address.BillingPhoneNumberIsInvalid`
+  * `Transaction.ShippingMethodIsInvalid`
+  * `Transaction.ShippingPhoneNumberIsInvalid`
+  * `Transaction.BillingPhoneNumberIsInvalid`
   * `RiskData.CustomerBrowserIsTooLong`
   * `RiskData.CustomerDeviceIdIsTooLong`
   * `RiskData.CustomerLocationZipInvalidCharacters`